requirements.txt
google-cloud-storage==2.14.0
gcsfs==2024.2.0
fsspec==2024.2.0
pandas==2.2.0
python-dotenv==1.0.1
tenacity==8.2.3
httpx==0.27.0